generate: add #define tags for enums with namespace-prefix
When using --renmae-policy 3 (alway), the tokens of the enum and the
target of the stable #define values were the same. This will cause the
preprocessor to replace all tokens in the enum, but with the same exact
value. No problem until we prefixed with namespace in the #define tags
in order to handle the enclosing namespace.

To make sure that the define-name is unique, prefix with 'ns'. This will
create unique labels and void the fields inside the enums to be expanded
from 'name' to 'namespace::name'.