modify: remove cornercase for namespace in asn1crt.h
namespace was added later due to the  extra logic at the astart of
asn1crt.h, however, due to dumb luck, the closing } worked well with the
closing 'extern "C" {', so the order sort of worked with older compiler.

With the new docker-build and newer asn1cc compiler, this triggered a
build-error, so clean this up and use from now on.